Added support for 'activeuser'
It will be possible to add activeuser as a datasource. The change uses lbs.limeDataconnection.ActiveUser.Record.
this will get the activeusers Record and make it available in the Viewmodel. Making it possible to compare an inspectors values against the coworker card of the actively logged in user.

The change will warn to the log if the activeuser record is missing. Tested briefly.
how to test:
Add datasource {type:'activeuser'} 
add <div data-bind="text:activeuser.name.text"></div> in the corresponding actionpad.